+++ database/migrations/2018_04_24_081209_create_user_profiles_table.php
@@ -24,7 +24,7 @@ class CreateUserProfilesTable extends Migration
                 ->references('id')
                 ->on('professions');
 
-            $table->unsignedInteger('user_id')->unique();
+            $table->unsignedBigInteger('user_id')->unique(); // UNSIGNED INT
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users')
